Fix loading and saving bugs and refactor files into packages
