+++ data/duke.txt
@@ -1,5 +1,6 @@
 T | 1 | read book
 D | 0 | return book | 2019-10-15
-E | 0 | project meeting | 2019-10-15 | 2019-10-16
+D | 0 | project meeting | 2019-10-15
 T | 1 | join sports club
 T | 0 | buy book
+E | 0 | buybook | 2019-10-15 | 2019-10-15
