--- src/main/java/Commands/Command.java
@@ -1,11 +1,0 @@
-package Commands;
-
-import Ui.Ui;
-import entities.*;
-
-import java.time.LocalDate;
-
-public abstract class Command {
-    public abstract void processCommand();
-
-}
