+++ src/main/java/Duke/Commands/Command.java
@@ -1,0 +1,6 @@
+package Duke.Commands;
+
+public abstract class Command {
+    public abstract void processCommand();
+
+}
