+++ src/main/java/Duke/Ui/Ui.java
@@ -1,4 +1,4 @@
-package Ui;
+package Duke.Ui;
 
 import java.util.Scanner;
 
@@ -12,7 +12,7 @@ public class Ui {
                 + "| |_| | |_| |   <  __/\n"
                 + "|____/ \\__,_|_|\\_\\___|\n";
         System.out.println("Hello from\n" + logo);
-        System.out.println("Hello! I'm Duke\nWhat can I do for you?");
+        System.out.println("Hello! I'm Duke.Duke\nWhat can I do for you?");
         printDivider();
     }
