+++ src/main/java/Duke/entities/Deadline.java
@@ -1,4 +1,4 @@
-package entities;
+package Duke.entities;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
@@ -13,6 +13,14 @@ public class Deadline extends Task {
 
     public LocalDate getBy() {
         return this.by;
+    }
+
+    @Override
+    public String toSave() {
+        if (super.isDone == true) {
+            return String.format("D | 1 | %s | %s\n", super.getDescription(), this.by);
+        }
+        return String.format("D | 0 | %s | %s\n", super.getDescription(), this.by);
     }
 
     @Override
