+++ src/main/java/Duke/entities/Event.java
@@ -1,4 +1,4 @@
-package entities;
+package Duke.entities;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
@@ -20,6 +20,15 @@ public class Event extends Task {
     public LocalDate getTo() {
         return this.to;
     }
+
+    @Override
+    public String toSave() {
+        if (super.isDone == true) {
+            return String.format("E | 1 | %s | %s | %s\n", super.getDescription(), this.from, this.to);
+        }
+        return String.format("E | 0 | %s | %s | %s\n", super.getDescription(), this.from, this.to);
+    }
+
 
     @Override
     public String toString() {
