+++ src/main/java/Duke/entities/Todo.java
@@ -1,4 +1,4 @@
-package entities;
+package Duke.entities;
 
 public class Todo extends Task {
     public Todo(String description) {
