+++ src/main/java/Duke/enums/CommandEnums.java
@@ -1,4 +1,4 @@
-package enums;
+package Duke.enums;
 
 public enum CommandEnums {
     BYE("bye"),
