+++ src/main/java/Duke/exceptions/InvalidInputException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package Duke.exceptions;
 
 public class InvalidInputException extends Exception {
     public InvalidInputException(String message) {
