+++ text-ui-test/EXPECTED.TXT
@@ -5,7 +5,7 @@ Hello from
 | |_| | |_| |   <  __/
 |____/ \__,_|_|\_\___|
 
-Hello! I'm Duke
+Hello! I'm Duke.Duke
 What can I do for you?
 ---------------------------------------
 alright, I've added the following task:
